requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
torch
jupyter
notebook
Pillow
PyMuPDF
